requirements.txt
attrs==19.3.0
Django==2.2.10
django-debug-toolbar==2.2
djangorestframework==3.11.0
Faker==0.9.1
importlib-metadata==1.5.0
mixer==6.1.3
more-itertools==8.2.0
mypy==0.761
mypy-extensions==0.4.3
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pyproj==2.4.2.post1
pytest==5.3.5
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.4.1
typing-extensions==3.7.4.1
wcwidth==0.1.8
zipp==3.0.0
